+++ docusaurus.config.ts
@@ -3,20 +3,27 @@ import type * as Preset from '@docusaurus/preset-classic';
 import {themes as prismThemes} from 'prism-react-renderer';
 import npm2yarn from '@docusaurus/remark-plugin-npm2yarn';
 import replace from './plugins/remark/replace-variables';
+const { ProvidePlugin } = require("webpack");
 
 const config: Config = {
   title: 'Reveal',
   tagline: 'Embedded Analytics & Business Intelligence Tools',
   url: 'https://help.revealbi.io',
   baseUrl: '/',
+  onBrokenAnchors: 'warn',
   onBrokenLinks: 'throw',
   onBrokenMarkdownLinks: 'throw',
   favicon: 'img/favicon.ico',
+  trailingSlash: true,
 
   // GitHub pages deployment config.
   // If you aren't using GitHub pages, you don't need these.
   organizationName: 'RevealBi', // Usually your GitHub org/user name.
   projectName: 'Documentation', // Usually your repo name.
+
+  plugins: [
+    ["docusaurus-node-polyfills", { onlyAliases: ["process"] }]
+  ],
 
   i18n: {
     defaultLocale: 'en',
@@ -61,6 +68,11 @@ const config: Config = {
   ],
 
   themeConfig: {
+    algolia: {
+      appId: 'QBVJ183IXZ',
+      apiKey: 'be244eaae98ce29feeb1882558494e62',
+      indexName: 'help-revealbi',
+    },
     navbar: {
       title: 'Reveal',
       logo: {